pyproject.toml
[tool.poetry]
name = "image-analysis-api"
version = "0.1.0"
description = "RESTful API for analyzing uploaded images"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
torchvision = "^0.15.0"
transformers = "^4.30.0"
opencv-python = "^4.7.0"
Pillow = "^10.0.0"
numpy = "^1.24.0"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.0"
loguru = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[tool.poetry.scripts]
start = "main:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
